Move structural validation code to `structural`
